Move deserialization into bridge/shared as well
