+++ rust/bridge/jni/src/lib.rs
@@ -104,8 +104,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_ECPublicKey_1Ver
     })
 }
 
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_ECPrivateKey_1Deserialize is PrivateKey::deserialize);
-
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_ECPrivateKey_1Serialize(PrivateKey) using
                        |k: &PrivateKey| Ok(k.serialize()));
 
@@ -284,8 +282,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_HKDF_1DeriveSecr
     })
 }
 
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SignalMessage_1Deserialize is SignalMessage::try_from);
-
 #[no_mangle]
 pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SignalMessage_1New(
     env: JNIEnv,
@@ -364,8 +360,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SignalMessage_1V
     })
 }
 
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_PreKeySignalMessage_1Deserialize is PreKeySignalMessage::try_from);
-
 #[no_mangle]
 pub unsafe extern "C" fn Java_org_signal_client_internal_Native_PreKeySignalMessage_1New(
     env: JNIEnv,
@@ -441,8 +435,6 @@ jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_PreKeySignalMessag
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_PreKeySignalMessage_1GetSerialized(PreKeySignalMessage) using
                        |m: &PreKeySignalMessage| Ok(m.serialized().to_vec()));
 
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SenderKeyMessage_1Deserialize is SenderKeyMessage::try_from);
-
 #[no_mangle]
 pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SenderKeyMessage_1New(
     env: JNIEnv,
@@ -462,8 +454,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SenderKeyMessage
         box_object::<SenderKeyMessage>(skm)
     })
 }
-
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SenderKeyDistributionMessage_1Deserialize is SenderKeyDistributionMessage::try_from);
 
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_SenderKeyMessage_1GetKeyId(SenderKeyMessage) using
                  |m: &SenderKeyMessage| Ok(m.key_id()));
@@ -634,8 +624,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SignedPreKeyReco
     })
 }
 
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SignedPreKeyRecord_1Deserialize is SignedPreKeyRecord::deserialize);
-
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_SignedPreKeyRecord_1GetId(SignedPreKeyRecord) using
                  SignedPreKeyRecord::id);
 
@@ -675,8 +663,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_PreKeyRecord_1Ne
         box_object::<PreKeyRecord>(Ok(pkr))
     })
 }
-
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_PreKeyRecord_1Deserialize is PreKeyRecord::deserialize);
 
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_PreKeyRecord_1GetId(PreKeyRecord) using
                  PreKeyRecord::id);
@@ -728,8 +714,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SenderKeyRecord_
         box_object::<SenderKeyRecord>(Ok(skr))
     })
 }
-
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SenderKeyRecord_1Deserialize is SenderKeyRecord::deserialize);
 
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_SenderKeyRecord_1GetSerialized(SenderKeyRecord) using
                        SenderKeyRecord::serialize);
@@ -1550,7 +1534,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_GroupCipher_1Dec
 }
 
 // SessionRecord
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SessionRecord_1Deserialize is SessionRecord::deserialize);
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_SessionRecord_1Serialize(SessionRecord) using SessionRecord::serialize);
 
 #[no_mangle]
@@ -1614,7 +1597,6 @@ jni_fn_get_jint!(Java_org_signal_client_internal_Native_SessionRecord_1GetSessio
 jni_fn_get_jboolean!(Java_org_signal_client_internal_Native_SessionRecord_1HasSenderChain(SessionRecord) using SessionRecord::has_sender_chain);
 
 // SessionState
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SessionState_1Deserialize is SessionState::deserialize);
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_SessionState_1GetSessionVersion(SessionState) using SessionState::session_version);
 jni_fn_get_jboolean!(Java_org_signal_client_internal_Native_SessionState_1HasSenderChain(SessionState) using SessionState::has_sender_chain);
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_SessionState_1Serialized(SessionState) using SessionState::serialize);
@@ -1736,8 +1718,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SessionRecord_1I
 }
 
 // Server Certificate
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_ServerCertificate_1Deserialize is ServerCertificate::deserialize);
-
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_ServerCertificate_1GetKeyId(ServerCertificate) using ServerCertificate::key_id);
 
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_ServerCertificate_1GetSerialized(ServerCertificate) using ServerCertificate::serialized);
@@ -1768,8 +1748,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_ServerCertificat
 }
 
 // Sender Certificate
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_SenderCertificate_1Deserialize is SenderCertificate::deserialize);
-
 jni_fn_get_jlong!(Java_org_signal_client_internal_Native_SenderCertificate_1GetExpiration(SenderCertificate) using SenderCertificate::expiration);
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_SenderCertificate_1GetDeviceId(SenderCertificate) using SenderCertificate::sender_device_id);
 
@@ -1868,8 +1846,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_SenderCertificat
 }
 
 // UnidentifiedSenderMessageContent
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessageContent_1Deserialize is UnidentifiedSenderMessageContent::deserialize);
-
 jni_fn_get_jint!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessageContent_1GetMsgType(UnidentifiedSenderMessageContent) using
                  |m: &UnidentifiedSenderMessageContent| Ok(m.msg_type()? as u32));
 
@@ -1906,8 +1882,6 @@ pub unsafe extern "C" fn Java_org_signal_client_internal_Native_UnidentifiedSend
 }
 
 // UnidentifiedSenderMessage
-jni_fn_deserialize!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessage_1Deserialize is UnidentifiedSenderMessage::deserialize);
-
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessage_1GetSerialized(UnidentifiedSenderMessage) using UnidentifiedSenderMessage::serialized);
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessage_1GetEncryptedMessage(UnidentifiedSenderMessage) using UnidentifiedSenderMessage::encrypted_message);
 jni_fn_get_jbytearray!(Java_org_signal_client_internal_Native_UnidentifiedSenderMessage_1GetEncryptedStatic(UnidentifiedSenderMessage) using UnidentifiedSenderMessage::encrypted_static);
