+++ rust/bridge/shared/Cargo.toml
@@ -15,7 +15,8 @@ libsignal-protocol-rust = { path = "../../protocol" }
 aes-gcm-siv = { path = "../../aes-gcm-siv" }
 paste = "1.0"
 
+libc = { version = "0.2", optional = true }
 jni = { version = "0.17", optional = true }
 
 [features]
-ffi = []
+ffi = ["libc"]
