+++ rust/bridge/shared/src/support_ffi/mod.rs
@@ -3,10 +3,14 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 //
 
+use libsignal_protocol_rust::*;
+
+pub(crate) use libc::{c_uchar, size_t};
+pub(crate) use paste::paste;
+pub(crate) use std::convert::TryFrom;
+
 mod error;
 pub use error::*;
-
-pub(crate) use paste::paste;
 
 pub fn run_ffi_safe<F: FnOnce() -> Result<(), SignalFfiError> + std::panic::UnwindSafe>(
     f: F,
@@ -20,6 +24,25 @@ pub fn run_ffi_safe<F: FnOnce() -> Result<(), SignalFfiError> + std::panic::Unwi
     match result {
         Ok(()) => std::ptr::null_mut(),
         Err(e) => Box::into_raw(Box::new(e)),
+    }
+}
+
+pub unsafe fn box_object<T>(
+    p: *mut *mut T,
+    obj: Result<T, SignalProtocolError>,
+) -> Result<(), SignalFfiError> {
+    if p.is_null() {
+        return Err(SignalFfiError::NullPointer);
+    }
+    match obj {
+        Ok(o) => {
+            *p = Box::into_raw(Box::new(o));
+            Ok(())
+        }
+        Err(e) => {
+            *p = std::ptr::null_mut();
+            Err(SignalFfiError::Signal(e))
+        }
     }
 }
 
@@ -42,4 +65,30 @@ macro_rules! bridge_destroy {
             bridge_destroy!($typ, ffi = [<$typ:snake>]);
         }
     }
+}
+
+macro_rules! bridge_deserialize {
+    ( $typ:ident::$fn:path, ffi = $ffi_name:ident $(, jni = $jni_name:ident)? ) => {
+        paste! {
+            #[no_mangle]
+            pub unsafe extern "C" fn [<signal_ $ffi_name _deserialize>](
+                p: *mut *mut $typ,
+                data: *const c_uchar,
+                data_len: size_t,
+            ) -> *mut SignalFfiError {
+                run_ffi_safe(|| {
+                    if data.is_null() {
+                        return Err(SignalFfiError::NullPointer);
+                    }
+                    let data = std::slice::from_raw_parts(data, data_len);
+                    box_object(p, $typ::$fn(data))
+                })
+            }
+        }
+    };
+    ( $typ:ident::$fn:path $(, jni = $jni_name:ident)? ) => {
+        paste! {
+            bridge_deserialize!($typ::$fn, ffi = [<$typ:snake>]);
+        }
+    };
 }
