+++ rust/bridge/shared/src/support_jni/error.rs
@@ -1,0 +1,91 @@
+//
+// Copyright 2020 Signal Messenger, LLC.
+// SPDX-License-Identifier: AGPL-3.0-only
+//
+
+use std::fmt;
+
+use aes_gcm_siv::Error as AesGcmSivError;
+use libsignal_protocol_rust::*;
+
+#[derive(Debug)]
+pub enum SignalJniError {
+    Signal(SignalProtocolError),
+    AesGcmSiv(AesGcmSivError),
+    Jni(jni::errors::Error),
+    BadJniParameter(&'static str),
+    UnexpectedJniResultType(&'static str, &'static str),
+    NullHandle,
+    IntegerOverflow(String),
+    UnexpectedPanic(std::boxed::Box<dyn std::any::Any + std::marker::Send>),
+    ExceptionDuringCallback(String),
+}
+
+impl SignalJniError {
+    pub fn to_signal_protocol_error(&self) -> SignalProtocolError {
+        match self {
+            SignalJniError::Signal(e) => e.clone(),
+            SignalJniError::Jni(e) => SignalProtocolError::FfiBindingError(e.to_string()),
+            SignalJniError::BadJniParameter(m) => {
+                SignalProtocolError::InvalidArgument(m.to_string())
+            }
+            _ => SignalProtocolError::FfiBindingError(format!("{}", self)),
+        }
+    }
+}
+
+impl fmt::Display for SignalJniError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        match self {
+            SignalJniError::Signal(s) => write!(f, "{}", s),
+            SignalJniError::AesGcmSiv(s) => write!(f, "{}", s),
+            SignalJniError::Jni(s) => write!(f, "JNI error {}", s),
+            SignalJniError::ExceptionDuringCallback(s) => {
+                write!(f, "exception recieved during callback {}", s)
+            }
+            SignalJniError::NullHandle => write!(f, "null handle"),
+            SignalJniError::BadJniParameter(m) => write!(f, "bad parameter type {}", m),
+            SignalJniError::UnexpectedJniResultType(m, t) => {
+                write!(f, "calling {} returned unexpected type {}", m, t)
+            }
+            SignalJniError::IntegerOverflow(m) => {
+                write!(f, "integer overflow during conversion of {}", m)
+            }
+            SignalJniError::UnexpectedPanic(e) => match e.downcast_ref::<&'static str>() {
+                Some(s) => write!(f, "unexpected panic: {}", s),
+                None => write!(f, "unknown unexpected panic"),
+            },
+        }
+    }
+}
+
+impl From<SignalProtocolError> for SignalJniError {
+    fn from(e: SignalProtocolError) -> SignalJniError {
+        SignalJniError::Signal(e)
+    }
+}
+
+impl From<AesGcmSivError> for SignalJniError {
+    fn from(e: AesGcmSivError) -> SignalJniError {
+        SignalJniError::AesGcmSiv(e)
+    }
+}
+
+impl From<jni::errors::Error> for SignalJniError {
+    fn from(e: jni::errors::Error) -> SignalJniError {
+        SignalJniError::Jni(e)
+    }
+}
+
+impl From<SignalJniError> for SignalProtocolError {
+    fn from(err: SignalJniError) -> SignalProtocolError {
+        match err {
+            SignalJniError::Signal(e) => e,
+            SignalJniError::Jni(e) => SignalProtocolError::FfiBindingError(e.to_string()),
+            SignalJniError::BadJniParameter(m) => {
+                SignalProtocolError::InvalidArgument(m.to_string())
+            }
+            _ => SignalProtocolError::FfiBindingError(format!("{}", err)),
+        }
+    }
+}
